Add a dedicated rotation command and keybinding
We need a dedicated command for rotation since when `pdf-tools`
redisplays the image, the `:rotation` image-property is lost.
Pdf-tools redisplays the images when changing pages, zooming in/out,
setting the slice, etc. so this is essential to have.

A variable is used simply because `pdf-tools` already uses a variable to
keep track of the current scaling.

* lisp/pdf-view.el (pdf-view--current-rotation): Add new variable to
keep track of angle.
(pdf-view-create-page): Account for above.
(pdf-view-mode-map): Add the new command under "R".
(pdf-view-rotate): Add new command.

Fixes: #152
Closes: #165